requirements/ci.txt
-r base.txt

pylint==3.3.2
pylint-django==2.6.1
black==24.10.0
isort==5.13.2
pytest==8.3.4
pytest-django==4.9.0
pytest-mock==3.14.0
pytest-cov==6.0.0
